# Improvements to SHA-1, SHA-256 and MD5 performance