--- a/IDE/ROWLEY-CROSSWORKS-ARM/test_main.c
+++ b/IDE/ROWLEY-CROSSWORKS-ARM/test_main.c
@@ -65,7 +65,7 @@ void main(void)
         test_num++;
     } while(args.return_code == 0);
 
-    /*Print this again for redundancy*/
+    /* Print this again for redundancy */
     #ifdef WOLFSSL_FRDM_K64_JENKINS
         printf("\n&&&&&&&&&&&&&& done &&&&&&&&&&&&&\n");
         delay_us(1000000);