--- a/wolfcrypt/src/sha.c
+++ b/wolfcrypt/src/sha.c
@@ -205,7 +205,9 @@
     #endif
 
     #define USE_SHA_SOFTWARE_IMPL /* Only for API's, actual transform is here */
-    #define XTRANSFORM(S,B)   Transform((S),(B))
+
+    #define XTRANSFORM(S,B)       Transform((S),(B))
+    #define XTRANSFORM_LEN(S,B,L) Transform_Len((S),(B),(L))
 
     static int InitSha(wc_Sha* sha)
     {
@@ -228,14 +230,29 @@
         return ret;
     }
 
-    static int Transform(wc_Sha* sha, byte* data)
+    static int Transform(wc_Sha* sha, const byte* data)
     {
         int ret = wolfSSL_CryptHwMutexLock();
         if(ret == 0) {
     #ifdef FREESCALE_MMCAU_CLASSIC_SHA
-            cau_sha1_hash_n(data, 1, sha->digest);
+            cau_sha1_hash_n((byte*)data, 1, sha->digest);
     #else
-            MMCAU_SHA1_HashN(data, 1, (uint32_t*)sha->digest);
+            MMCAU_SHA1_HashN((byte*)data, 1, (uint32_t*)sha->digest);
+    #endif
+            wolfSSL_CryptHwMutexUnLock();
+        }
+        return ret;
+    }
+
+    static int Transform_Len(wc_Sha* sha, const byte* data, word32 len)
+    {
+        int ret = wolfSSL_CryptHwMutexLock();
+        if(ret == 0) {
+    #ifdef FREESCALE_MMCAU_CLASSIC_SHA
+            cau_sha1_hash_n((byte*)data, len/WC_SHA_BLOCK_SIZE, sha->digest);
+    #else
+            MMCAU_SHA1_HashN((byte*)data, len/WC_SHA_BLOCK_SIZE,
+                (uint32_t*)sha->digest);
     #endif
             wolfSSL_CryptHwMutexUnLock();
         }
@@ -280,12 +297,12 @@
 
         return ret;
     }
-    
+
 #elif defined(WOLFSSL_RENESAS_TSIP_CRYPT) && \
     !defined(NO_WOLFSSL_RENESAS_TSIP_CRYPT_HASH)
-    
+
     /* implemented in wolfcrypt/src/port/Renesas/renesas_tsip_sha.c */
-    
+
 #else
     /* Software implementation */
     #define USE_SHA_SOFTWARE_IMPL
@@ -312,6 +329,10 @@
 
 #endif /* End Hardware Acceleration */
 
+#ifndef WC_SHA_DATA_ALIGNMENT
+    /* default to 32-bit alignement */
+    #define WC_SHA_DATA_ALIGNMENT 4
+#endif
 
 /* Software implementation */
 #ifdef USE_SHA_SOFTWARE_IMPL
@@ -327,7 +348,7 @@ static WC_INLINE void AddLength(wc_Sha* sha, word32 len)
 #ifndef XTRANSFORM
     #define XTRANSFORM(S,B)   Transform((S),(B))
 
-    #define blk0(i) (W[i] = sha->buffer[i])
+    #define blk0(i) (W[i] = *((word32*)&data[i*sizeof(word32)]))
     #define blk1(i) (W[(i)&15] = \
         rotlFixed(W[((i)+13)&15]^W[((i)+8)&15]^W[((i)+2)&15]^W[(i)&15],1))
 
@@ -356,7 +377,7 @@ static WC_INLINE void AddLength(wc_Sha* sha, word32 len)
     #define R4(v,w,x,y,z,i) (z)+= f4((w),(x),(y)) + blk1((i)) + 0xCA62C1D6+ \
         rotlFixed((v),5); (w) = rotlFixed((w),30);
 
-    static void Transform(wc_Sha* sha, byte* data)
+    static int Transform(wc_Sha* sha, const byte* data)
     {
         word32 W[WC_SHA_BLOCK_SIZE / sizeof(word32)];
 
@@ -431,6 +452,8 @@ static WC_INLINE void AddLength(wc_Sha* sha, word32 len)
         sha->digest[4] += e;
 
         (void)data; /* Not used */
+
+        return 0;
     }
 #endif /* !USE_CUSTOM_SHA_TRANSFORM */
 
@@ -466,16 +489,17 @@ int wc_InitSha_ex(wc_Sha* sha, void* heap, int devId)
     return ret;
 }
 
+/* do block size increments/updates */
 int wc_ShaUpdate(wc_Sha* sha, const byte* data, word32 len)
 {
+    int ret = 0;
+    word32 blocksLen;
     byte* local;
+    word32* local32;
 
-    if (sha == NULL ||(data == NULL && len > 0)) {
+    if (sha == NULL || (data == NULL && len > 0)) {
         return BAD_FUNC_ARG;
     }
-
-    /* do block size increments */
-    local = (byte*)sha->buffer;
 
 #ifdef WOLF_CRYPTO_CB
     if (sha->devId != INVALID_DEVID) {
@@ -497,37 +521,107 @@ int wc_ShaUpdate(wc_Sha* sha, const byte* data, word32 len)
     if (sha->buffLen >= WC_SHA_BLOCK_SIZE)
         return BUFFER_E;
 
-    while (len) {
-        word32 add = min(len, WC_SHA_BLOCK_SIZE - sha->buffLen);
-        XMEMCPY(&local[sha->buffLen], data, add);
+    if (data == NULL && len == 0) {
+        /* valid, but do nothing */
+        return 0;
+    }
 
-        sha->buffLen += add;
-        data         += add;
-        len          -= add;
+    /* add length for final */
+    AddLength(sha, len);
+
+    local = (byte*)sha->buffer;
+    local32 = sha->buffer;
+
+    /* process any remainder from previous operation */
+    if (sha->buffLen > 0) {
+        blocksLen = min(len, WC_SHA_BLOCK_SIZE - sha->buffLen);
+        XMEMCPY(&local[sha->buffLen], data, blocksLen);
+
+        sha->buffLen += blocksLen;
+        data         += blocksLen;
+        len          -= blocksLen;
 
         if (sha->buffLen == WC_SHA_BLOCK_SIZE) {
-#if defined(LITTLE_ENDIAN_ORDER) && !defined(FREESCALE_MMCAU_SHA)
-            ByteReverseWords(sha->buffer, sha->buffer, WC_SHA_BLOCK_SIZE);
-#endif
-#if !defined(WOLFSSL_ESP32WROOM32_CRYPT) || \
-    defined(NO_WOLFSSL_ESP32WROOM32_CRYPT_HASH)
-            XTRANSFORM(sha, local);
-#else
-            if(sha->ctx.mode == ESP32_SHA_INIT){
+        #if defined(LITTLE_ENDIAN_ORDER) && !defined(FREESCALE_MMCAU_SHA)
+            ByteReverseWords(local32, local32, WC_SHA_BLOCK_SIZE);
+        #endif
+
+        #if defined(WOLFSSL_ESP32WROOM32_CRYPT) && \
+            !defined(NO_WOLFSSL_ESP32WROOM32_CRYPT_HASH)
+            if (sha->ctx.mode == ESP32_SHA_INIT) {
                 esp_sha_try_hw_lock(&sha->ctx);
             }
-            if(sha->ctx.mode == ESP32_SHA_SW){
-                XTRANSFORM(sha, local);
+            if (sha->ctx.mode == ESP32_SHA_SW {
+                ret = XTRANSFORM(sha, (const byte*)local);
             } else {
-                esp_sha_process(sha);
+                esp_sha_process(sha, (const byte*)local);
             }
-#endif
-            AddLength(sha, WC_SHA_BLOCK_SIZE);
+        #else
+            ret = XTRANSFORM(sha, (const byte*)local);
+        #endif
+            if (ret != 0)
+                return ret;
+
             sha->buffLen = 0;
         }
     }
 
-    return 0;
+    /* process blocks */
+#ifdef XTRANSFORM_LEN
+    /* get number of blocks */
+    /* 64-1 = 0x3F (~ Inverted = 0xFFFFFFC0) */
+    /* len (masked by 0xFFFFFFC0) returns block aligned length */
+    blocksLen = len & ~(WC_SHA_BLOCK_SIZE-1);
+    if (blocksLen > 0) {
+        /* Byte reversal performed in function if required. */
+        XTRANSFORM_LEN(sha, data, blocksLen);
+        data += blocksLen;
+        len  -= blocksLen;
+    }
+#else
+    while (len >= WC_SHA_BLOCK_SIZE) {
+        /* optimization to avoid memcpy if data pointer is properly aligned */
+        /* Little Endian requires byte swap, so can't use data directly */
+    #if defined(WC_SHA_DATA_ALIGNMENT) && !defined(LITTLE_ENDIAN_ORDER)
+        if (((size_t)data % WC_SHA_DATA_ALIGNMENT) == 0) {
+            local32 = (word32*)data;
+        }
+        else
+    #endif
+        {
+            XMEMCPY(local32, data, WC_SHA_BLOCK_SIZE);
+        }
+
+        data += WC_SHA_BLOCK_SIZE;
+        len  -= WC_SHA_BLOCK_SIZE;
+
+    #if defined(LITTLE_ENDIAN_ORDER) && !defined(FREESCALE_MMCAU_SHA)
+        ByteReverseWords(local32, local32, WC_SHA_BLOCK_SIZE);
+    #endif
+
+    #if defined(WOLFSSL_ESP32WROOM32_CRYPT) && \
+        !defined(NO_WOLFSSL_ESP32WROOM32_CRYPT_HASH)
+        if (sha->ctx.mode == ESP32_SHA_INIT){
+            esp_sha_try_hw_lock(&sha->ctx);
+        }
+        if (sha->ctx.mode == ESP32_SHA_SW){
+            ret = XTRANSFORM(sha, (const byte*)local32);
+        } else {
+            esp_sha_process(sha, (const byte*)local32);
+        }
+    #else
+        ret = XTRANSFORM(sha, (const byte*)local32);
+    #endif
+    }
+#endif /* XTRANSFORM_LEN */
+
+    /* save remainder */
+    if (len > 0) {
+        XMEMCPY(local, data, len);
+        sha->buffLen = len;
+    }
+
+    return ret;
 }
 
 int wc_ShaFinalRaw(wc_Sha* sha, byte* hash)
@@ -552,6 +646,7 @@ int wc_ShaFinalRaw(wc_Sha* sha, byte* hash)
 
 int wc_ShaFinal(wc_Sha* sha, byte* hash)
 {
+    int ret;
     byte* local;
 
     if (sha == NULL || hash == NULL) {
@@ -576,8 +671,6 @@ int wc_ShaFinal(wc_Sha* sha, byte* hash)
     }
 #endif /* WOLFSSL_ASYNC_CRYPT */
 
-    AddLength(sha, sha->buffLen);  /* before adding pads */
-
     local[sha->buffLen++] = 0x80;  /* add 1 */
 
     /* pad with zeros */
@@ -585,22 +678,26 @@ int wc_ShaFinal(wc_Sha* sha, byte* hash)
         XMEMSET(&local[sha->buffLen], 0, WC_SHA_BLOCK_SIZE - sha->buffLen);
         sha->buffLen += WC_SHA_BLOCK_SIZE - sha->buffLen;
 
-#if defined(LITTLE_ENDIAN_ORDER) && !defined(FREESCALE_MMCAU_SHA)
+    #if defined(LITTLE_ENDIAN_ORDER) && !defined(FREESCALE_MMCAU_SHA)
         ByteReverseWords(sha->buffer, sha->buffer, WC_SHA_BLOCK_SIZE);
-#endif
-#if !defined(WOLFSSL_ESP32WROOM32_CRYPT) || \
-    defined(NO_WOLFSSL_ESP32WROOM32_CRYPT_HASH)
-        XTRANSFORM(sha, local);
-#else
-        if(sha->ctx.mode == ESP32_SHA_INIT){
+    #endif
+
+    #if defined(WOLFSSL_ESP32WROOM32_CRYPT) && \
+        !defined(NO_WOLFSSL_ESP32WROOM32_CRYPT_HASH)
+        if (sha->ctx.mode == ESP32_SHA_INIT) {
             esp_sha_try_hw_lock(&sha->ctx);
         }
-        if(sha->ctx.mode == ESP32_SHA_SW){
-            XTRANSFORM(sha, local);
+        if (sha->ctx.mode == ESP32_SHA_SW) {
+            ret = XTRANSFORM(sha, (const byte*)local);
         } else {
-            esp_sha_process(sha);
+            esp_sha_process(sha, (const byte*)local);
         }
-#endif
+    #else
+        ret = XTRANSFORM(sha, (const byte*)local);
+    #endif
+        if (ret != 0)
+            return ret;
+
         sha->buffLen = 0;
     }
     XMEMSET(&local[sha->buffLen], 0, WC_SHA_PAD_SIZE - sha->buffLen);
@@ -625,26 +722,29 @@ int wc_ShaFinal(wc_Sha* sha, byte* hash)
                      2 * sizeof(word32));
 #endif
 
-#if !defined(WOLFSSL_ESP32WROOM32_CRYPT) || \
-    defined(NO_WOLFSSL_ESP32WROOM32_CRYPT_HASH)
-    XTRANSFORM(sha, local);
-#else
-    if(sha->ctx.mode == ESP32_SHA_INIT){
+#if defined(WOLFSSL_ESP32WROOM32_CRYPT) && \
+    !defined(NO_WOLFSSL_ESP32WROOM32_CRYPT_HASH)
+    if (sha->ctx.mode == ESP32_SHA_INIT) {
         esp_sha_try_hw_lock(&sha->ctx);
     }
-    if(sha->ctx.mode == ESP32_SHA_SW){
-        XTRANSFORM(sha, local);
+    if (sha->ctx.mode == ESP32_SHA_SW) {
+        ret = XTRANSFORM(sha, (const byte*)local);
     } else {
         esp_sha_digest_process(sha, 1);
     }
+#else
+    ret = XTRANSFORM(sha, (const byte*)local);
 #endif
 
 #ifdef LITTLE_ENDIAN_ORDER
     ByteReverseWords(sha->digest, sha->digest, WC_SHA_DIGEST_SIZE);
 #endif
+
     XMEMCPY(hash, sha->digest, WC_SHA_DIGEST_SIZE);
 
-    return InitSha(sha); /* reset state */
+    (void)InitSha(sha); /* reset state */
+
+    return ret;
 }
 
 #endif /* USE_SHA_SOFTWARE_IMPL */
@@ -707,7 +807,7 @@ int wc_ShaGetHash(wc_Sha* sha, byte* hash)
         sha->ctx.mode = ESP32_SHA_SW;
 #endif
 
-    
+
     }
     return ret;
 }